Add filters
Filters are a way to trim down the amount of log lines that are
going to be passed to commands for further investigation.

Filters are meant to be used as a way to focus the commands output
to certain specific log lines, i.e. only run commands on log lines
with a specific IP, on SSL connections or that contains a path...