Improve performance of cmd_print
My log file has about 25k log entries after filtering. With the previous
implementation, `cmd_print` hangs for a long time with CPU at 100%.

Using for loop and concate a string is expensive, since the previous
`data` object is thrown away and may get GC'ed.

Switched to a list comprehension and this function finished in 0.01
second (on pypy, haven't tested it on cpython).